experience level change from float to int

+++ core/src/com/mygdx/game/GameObjekts/SpaceShipParts/FuelTank.java
@@ -3,7 +3,7 @@ package com.mygdx.game.GameObjekts.SpaceShipParts;
 public class FuelTank extends ShipModule{
 
 
-    public FuelTank(ModuleType moduleType, String name, double capacity, double fill, double cost, double baseFailureRate){
+    public FuelTank(ModuleType moduleType, String name, double capacity, double fill, double cost, int baseFailureRate){
         super(moduleType, name, capacity, cost);
         this.capacity = capacity;
         this.fill = fill;
